mrg utils + openai
